portability fix for Strophe.Builder.prototype.cnode()
The cnode() method makes use of the DOM appendChild(elem) method
which is only guaranteed to work if "elem" originates from the
same DOM document. While some browsers such as Firefox also accept
foreign DOM elements here, other browsers like Google Chrome stick
to the standard and don't accept those.

The correct way to solve this issue (as intended by the standard)
is to import the foreign DOM element via importNode() before
appending it. However, importNode() is a method of the DOM document,
not the element. Since the DOM document is not directly available
to Strophe.Builder, but somewhat hidden in Strophe in an internal
"_xmlGenerator" attribute, this would require a new helper function
like "Strophe.importElement()".

This patch avoids adding such a method by using "Strophe.copyElement()"
instead. This function works as well because it completely rebuilds
the DOM structure using Strophe's internal DOM document.

This makes the patch very simple and short, but might not lead to
optimal performance.