SASL OAUTHBEARER should not require a JID node.
When performing authentication via the SASL OAUTHBEARER mechanism, the provided token
often contains the desired user identity. As the OAUTHBEARER mechanism specification
does not _require_ this, such an identity can, optionally, be provided as a value
distinct from the token.

In the old implementation of OAUTHBEARER in Strophe.js, the node-part of the JID provided
to the connect method was used for this optional user identity. Strophe.js regarded the
node-part as a required attribute for OAUTHBEARER, and would always use its value.

When the access token contains the desired user identity, it is undesirable to explicitly
have to define this identity 'out-of-band' of the token again. This would require the
client to be able to parse the token, or have knowledge of the desired identity in another
fashion.

Instead, it is desirable to be able to connect with only the access token, when this token
contains the desired user identity. When needed, a user identity can be provided via the
node-part of the JID, but this should be optional.