Add `clientChallenge` SASL mechanism method
This is to prevent the `onChallenge` method from being monkey patched for the
SCRAM-SHA1 auth mechanism, which causes issues when reconnecting.

An alternative fix would be to call `registerSASLMechanisms` in the
`reset` method to undo the patch, but I've opted for this approach since
the patch itself is a code smell.